rename uppercase JadeX dir to lowercase jadey
