--- test/fixtures/jadex/options.json
@@ -1,4 +1,0 @@
-{
-  "sourceType": "module",
-  "plugins": ["jadex", "flow"]
-}
