+++ test/fixtures/jadey/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["jadey", "flow"]
+}
